+++ src/pages/inventory/layoutPage/EmployeeInventoryDashboardLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from '../../../components/shared/AdminHeader'
+import Header from '../../../components/shared/Header'
 import SideBar from '../../../components/shared/SideBar'
 import '../../../styles/inventory/inventoryLayout/inventoryLayout.css'
 import EmployeeInventoryDashboard from '../pages/EmployeeInventoryDashboard'
