+++ src/pages/timeEntry/EmployeeAttendanceEntryPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
-import Header from '../../components/shared/AdminHeader'
+import Header from '../../components/shared/Header'
 import SideBar from '../../components/shared/SideBar'
 import TimeEntryContainer from '../../components/timeEntry/EmployeeAtendanceEntry/TimeEntryContainer'
 import '../../styles/timeEntry/pages/EmployeeAttendanceEntryPage.css'
