+++ src/pages/timeEntry/ManagerDashBoardPage.tsx
@@ -6,7 +6,7 @@ import Card from '../../components/timeEntry/Dashboard/Card'
 import { ReactComponent as AttendanceIcon } from '../../icons/timeEntry/AttendanceIcon.svg'
 import { ReactComponent as CalendarPlusIcon } from '../../icons/timeEntry/CalendarPlusIcon.svg'
 import { ReactComponent as RecordIcon } from '../../icons/timeEntry/RecordIcon.svg'
-import Header from '../../components/shared/AdminHeader'
+import Header from '../../components/shared/Header'
 import SideBar from '../../components/shared/SideBar'
 import EmployeeAttendancePieChart from '../../components/timeEntry/Dashboard/EmployeeAttendancePieChart'
 import EmployeeCalendar from '../../components/timeEntry/Dashboard/EmployeeCalendar'
